+++ cpp/libjoynr/include/joynr/ProxyBuilder.h
@@ -37,6 +37,7 @@
 #include "joynr/Future.h"
 #include "joynr/PrivateCopyAssign.h"
 #include "joynr/IProxyBuilder.h"
+#include "joynr/Logger.h"
 
 namespace joynr
 {
@@ -123,6 +124,8 @@ private:
     std::shared_ptr<MessageRouter> messageRouter;
     std::uint64_t messagingMaximumTtlMs;
     DiscoveryQos discoveryQos;
+
+    ADD_LOGGER(ProxyBuilder);
 };
 
 template <class T>
@@ -198,7 +201,15 @@ void ProxyBuilder<T>::buildAsync(
             return;
         }
 
+        JOYNR_LOG_DEBUG(logger,
+                        "DISCOVERY proxy created for provider participantId: {}, for domain: [{}], "
+                        "interface: {}",
+                        participantId,
+                        domain,
+                        T::INTERFACE_NAME());
+
         bool useInProcessConnector = requestCallerDirectory->containsRequestCaller(participantId);
+
         std::unique_ptr<T> proxy(proxyFactory.createProxy<T>(domain, messagingQos));
         proxy->handleArbitrationFinished(participantId, useInProcessConnector);
 
@@ -231,6 +242,9 @@ ProxyBuilder<T>* ProxyBuilder<T>::setDiscoveryQos(const DiscoveryQos& discoveryQ
     this->discoveryQos = discoveryQos;
     return this;
 }
+
+template <class T>
+INIT_LOGGER(ProxyBuilder<T>);
 
 } // namespace joynr
 #endif // PROXYBUILDER_H